% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alr_trans.R
\name{alr_trans}
\alias{alr_trans}
\title{The additive log-ratio transformation}
\usage{
alr_trans(eBay.comps)
}
\arguments{
\item{eBay.comps}{the estimated microbial compositions by empirical Bayes.}
}
\value{
Returns the alr transfromation of microbial compositions.
}
\description{
The additive log-ratio transformation
}
\examples{
library(phyloseq)
otu_tab <- t(combo.phyloseq.obj@otu_table@.Data)
tree <- phy_tree(combo.phyloseq.obj)
eBay.comps <- eBay_comps(otu_tab,prior = "Dirichlet", tree=NULL)
eBay.comps.alr <- alr_trans(eBay.comps)
}
