% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clr_trans.R
\name{clr_trans}
\alias{clr_trans}
\title{The centered log-ratio transformation}
\usage{
clr_trans(eBay.comps)
}
\arguments{
\item{eBay.comps}{the estimated microbial compositions by empirical Bayes.}
}
\value{
Returns the clr transfromation of microbial compositions.
}
\description{
The centered log-ratio transformation
}
\examples{
library(phyloseq)
otu_tab <- t(combo.phyloseq.obj@otu_table@.Data)
tree <- phy_tree(combo.phyloseq.obj)
eBay.comps <- eBay_comps(otu_tab,prior = "Dirichlet", tree=NULL)
eBay.comps.clr <- clr_trans(eBay.comps)
}
