% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract_BIC.R
\name{Extract_BIC}
\alias{Extract_BIC}
\title{Calculates the Bayesian information criterion for a list of MGLM objects.}
\usage{
Extract_BIC(x)
}
\arguments{
\item{x}{A list of MGLM objects. 'MGLMdtmFit', 'MGLMdtmReg', 'MGLMdtmSparseReg', or 'MGLMdtmTune'.}
}
\value{
Returns the BIC of a MGLMdtm object (a list of MGLM objects).
}
\description{
This function extracts BIC from a list of MGLM objects.
}
\examples{
library(phyloseq)
otu.tab <- t(otu_table(combo.phyloseq.obj))
tree <- phy_tree(combo.phyloseq.obj)
fit <- MGLMdtmFit(otu.tab, tree)
Extract_BIC(fit)
}
