% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract_logL.R
\name{Extract_logL}
\alias{Extract_logL}
\title{Calculates the loglikelihood for a list of MGLM objects.}
\usage{
Extract_logL(x)
}
\arguments{
\item{x}{A list of MGLM objects. 'MGLMdtmFit', 'MGLMdtmReg', 'MGLMdtmSparseReg', or 'MGLMdtmTune'.}
}
\value{
Returns the loglikelihood of a MGLMdtm object (a list of MGLM objects).
}
\description{
This function extracts loglikelihood from a list of MGLM objects.
}
\examples{
library(phyloseq)
otu.tab <- t(otu_table(combo.phyloseq.obj))
tree <- phy_tree(combo.phyloseq.obj)
fit <- MGLMdtmFit(otu.tab, tree)
Extract_logL(fit)
}
