% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICgenlasso.R
\name{ICgenlasso}
\alias{ICgenlasso}
\title{Calculates the Bayesian information criterion or Akaike's information criterion for a genlasso object.}
\usage{
ICgenlasso(fit, IC = "BIC")
}
\arguments{
\item{fit}{a genlasso object.}

\item{IC}{IC = BIC for Bayesian information criterion and IC = AIC for Akaike's information criterion.}
}
\value{
Returns the minimal AIC or BIC and the corresponding lambda.
}
\description{
This function extracts BIC or AIC from a genlasso object.
}
\examples{
library(phyloseq)
otu.tab <- t(otu_table(combo.phyloseq.obj))
X <- eBay_comps(otu.tab, prior = "Dirichlet")
tree <- phy_tree(combo.phyloseq.obj)
metadata <- sample_data(combo.phyloseq.obj)
y <- metadata$bmi
fit <- TASSO(y, X, tree)
IC <- ICgenlasso(fit, IC="AIC")
IC <- ICgenlasso(fit, IC="BIC")
}
