% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGLMdtmFit.R
\name{MGLMdtmFit}
\alias{MGLMdtmFit}
\title{Fit the Dirichlet-tree multinomial (DTM) distribution.}
\usage{
MGLMdtmFit(otu.tab, tree)
}
\arguments{
\item{otu.tab}{a data frame or matrix containing the count data. Rows of the matrix represent observations and columns are the taxa.}

\item{tree}{the phylogenetic tree.}
}
\value{
Returns a list of MGLM objects ('MGLMfit') corresponding to internal nodes of the tree.
}
\description{
Fit the Dirichlet-tree multinomial (DTM) distribution.
}
\examples{
library(phyloseq)
otu.tab <- t(otu_table(combo.phyloseq.obj))
tree <- phy_tree(combo.phyloseq.obj)
fit <- MGLMdtmFit(otu.tab, tree)
Extract_logL(fit)
}
\references{
Wang, T., Zhao, H.: A Dirichlet-tree multinomial regression model for associating dietary nutrients with gut microorganisms. Biometrics 73(3), 792–801 (2017).
}
