% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGLMdtmReg.R
\name{MGLMdtmReg}
\alias{MGLMdtmReg}
\title{Fit the Dirichlet-tree multinomial (DTM) regression.}
\usage{
MGLMdtmReg(otu.tab, X, tree)
}
\arguments{
\item{otu.tab}{a data frame or matrix containing the count data. Rows of the matrix represent observations and columns are the taxa.}

\item{X}{a data frame or matrix containing the covariates. Rows of the matrix represent observations.}

\item{tree}{the phylogenetic tree.}
}
\value{
Returns a list of MGLM objects ('MGLMreg') corresponding to internal nodes of the tree.
}
\description{
Fit the Dirichlet-tree multinomial (DTM) regression.
}
\examples{
library(phyloseq)
otu.tab <- t(otu_table(combo.phyloseq.obj))
tree <- phy_tree(combo.phyloseq.obj)
sodium <- sample_data(combo.phyloseq.obj)$sodium
fit <- MGLMdtmReg(otu.tab, sodium, tree)
Extract_logL(fit)
}
\references{
Wang, T., Zhao, H.: A Dirichlet-tree multinomial regression model for associating dietary nutrients with gut microorganisms. Biometrics 73(3), 792–801 (2017).
}
