% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGLMdtmSparseReg.R
\name{MGLMdtmSparseReg}
\alias{MGLMdtmSparseReg}
\title{Fit the Dirichlet-tree multinomial (DTM) sparse regression.}
\usage{
MGLMdtmSparseReg(otu.tab, X, tree, penalty = "sweep", lambda = Inf)
}
\arguments{
\item{otu.tab}{a data frame or matrix containing the count data. Rows of the matrix represent observations and columns are the taxa.}

\item{X}{a data frame or matrix containing the covariates. Rows of the matrix represent observations.}

\item{tree}{the phylogenetic tree.}

\item{penalty}{penalty type for the regularization term. Can be chosen from 'sweep', 'group', or 'nuclear'. See Details for the description of each penalty type of MGLM.}

\item{lambda}{penalty parameter.}
}
\value{
Returns a list of MGLM objects ('MGLMsparsereg') corresponding to internal nodes of the tree.
}
\description{
Fit the Dirichlet-tree multinomial (DTM) sparse regression.
}
\examples{
library(phyloseq)
library(magrittr)
otu.tab <- t(otu_table(combo.phyloseq.obj))
metadata <- sample_data(combo.phyloseq.obj)
tree <- phy_tree(combo.phyloseq.obj)
X <- as.matrix(metadata)[,18:37] \%>\% apply(., 2, as.numeric)
fit <- MGLMdtmSparseReg(otu.tab, X, tree, penalty = 'sweep', lambda = Inf)
Extract_logL(fit)
}
\references{
Wang, T., Zhao, H.: A Dirichlet-tree multinomial regression model for associating dietary nutrients with gut microorganisms. Biometrics 73(3), 792–801 (2017).
}
