% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_onTree.R
\name{plot_onTree}
\alias{plot_onTree}
\title{Visualization on the tree}
\usage{
plot_onTree(
  results,
  tree,
  model = "TASSO",
  IC = "BIC",
  tree.type = "fan",
  color = "darkred"
)
}
\arguments{
\item{results}{results from \link{TASSO} or \link{TreeFusedlasso}.}

\item{tree}{the phylogenetic tree.}

\item{model}{indicate the model used.}

\item{IC}{IC = BIC for Bayesian information criterion and IC = AIC for Akaike's information criterion.}

\item{tree.type}{a character string specifying the type of phylogeny to be drawn; it must be one of "phylogram" (the default), "cladogram", "fan", "unrooted", "radial" or any unambiguous abbreviation of these.}

\item{color}{the colours used for the tip labels or choosed pattern.}
}
\value{
A ggplot object.
}
\description{
The results of TASSO and TreeFusedlasso vasualized on the tree.
}
\examples{
library(phyloseq)
otu.tab <- t(otu_table(combo.phyloseq.obj))
eBay.comps <- eBay_comps(otu.tab, prior = "Dirichlet")
tree <- phy_tree(combo.phyloseq.obj)
metadata <- sample_data(combo.phyloseq.obj)
y <- metadata$bmi
fit_tasso <- TASSO(y, eBay.comps, tree)
plot_onTree(fit_tasso, tree)

fit_tflasso1 <- TreeFusedlasso(y, eBay.comps, tree)
plot_onTree(fit_tflasso1, tree, model='TreeFusedlasso')
}
