% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TASSO.R
\name{TASSO}
\alias{TASSO}
\title{Fit the Tree-guided Automatic Sub-composition Selection Operator (TASSO) model.}
\usage{
TASSO(y, X, tree = NULL)
}
\arguments{
\item{y}{the univariate response.}

\item{X}{a data frame or matrix containing the compositional data. Rows of the matrix represent observations and columns are the taxa.}

\item{tree}{when tree=NULL, it reduces to constrained lasso.}
}
\value{
Returns a genlasso object.
}
\description{
Fit the Tree-guided Automatic Sub-composition Selection Operator (TASSO) model.
}
\examples{
library(phyloseq)
otu.tab <- t(otu_table(combo.phyloseq.obj))
X <- eBay_comps(otu.tab, prior = "Dirichlet")
tree <- phy_tree(combo.phyloseq.obj)
metadata <- sample_data(combo.phyloseq.obj)
y <- metadata$bmi
fit <- TASSO(y, X, tree)
IC <- ICgenlasso(fit, IC='AIC')
IC <- ICgenlasso(fit, IC='BIC')
}
\references{
Wang, T., Zhao, H.: Structured subcomposition selection in regression and its application to microbiome data analysis. The Annals of Applied Statistics 11(2), 771–791 (2017).
}
