% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeFusedlasso.R
\name{TreeFusedlasso}
\alias{TreeFusedlasso}
\title{Fit the tree-guided fused lasso model.}
\usage{
TreeFusedlasso(y, X, tree, type = 1)
}
\arguments{
\item{y}{the univariate response.}

\item{X}{a data frame or matrix containing the compositional data. Rows of the matrix represent observations and columns are the taxa.}

\item{tree}{the phylogenetic tree.}

\item{type}{two different penalty construct methods : 1 or 2.}
}
\value{
Returns a genlasso object.
}
\description{
Fit the tree-guided fused lasso model.
}
\examples{
library(phyloseq)
otu.tab <- t(otu_table(combo.phyloseq.obj))
X <- eBay_comps(otu.tab, prior = "Dirichlet")
tree <- phy_tree(combo.phyloseq.obj)
metadata <- sample_data(combo.phyloseq.obj)
y <- metadata$bmi
fit <- TreeFusedlasso(y, X, tree)
IC <- ICgenlasso(fit, IC='AIC')
IC <- ICgenlasso(fit, IC='BIC')
}
\references{
Wang, T., Zhao, H.: Constructing predictive microbial signatures at multiple taxonomic levels. Journal of the American Statistical Association 112(519), 1022–1031 (2017).
}
