% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YtreeFun.R
\name{YtreeFun}
\alias{YtreeFun}
\title{Calculate the count on each subtree.}
\usage{
YtreeFun(otu.tab, tree)
}
\arguments{
\item{otu.tab}{a data frame or matrix containing the count data. Rows of the matrix represent observations and columns are the taxa.}

\item{tree}{the phylogenetic tree.}
}
\value{
Returns a list of abundance matrices corresponding to internal nodes of the tree.
}
\description{
Calculate the count on each subtree.
}
\examples{
library(phyloseq)
otu.tab <- t(otu_table(combo.phyloseq.obj))
tree <- phy_tree(combo.phyloseq.obj)
ytree <- YtreeFun(otu.tab, tree)
length(ytree)
}
