% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIdtmFit.R
\name{ZIdtmFit}
\alias{ZIdtmFit}
\title{Fit the Zero-inflated Dirichlet-tree multinomial (ZIDTM) distribution.}
\usage{
ZIdtmFit(otu.tab, tree)
}
\arguments{
\item{otu.tab}{a data frame or matrix containing the count data. Rows of the matrix represent observations and columns are the taxa.}

\item{tree}{the phylogenetic tree (should be binary here).}
}
\value{
Returns a list, containing results of each subtree.
}
\description{
Fit the Zero-inflated Dirichlet-tree multinomial (ZIDTM) distribution.
}
\examples{
library(phyloseq)
data("combo.phyloseq.obj")
otu.tab <- t(otu_table(combo.phyloseq.obj))
tree <- phy_tree(combo.phyloseq.obj)
fit <- ZIdtmFit(otu.tab, tree)
Extract_logL(fit)
}
\references{
Zhou, C., Zhao, H., and Wang, T.: Transformation and differential abundance analysis of microbiome data incorporating phylogeny. Bioinformatics btab543, https://doi.org/10.1093/bioinformatics/btab543 (2021).
}
