% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIdtmReg.R
\name{ZIdtmReg}
\alias{ZIdtmReg}
\title{Fit the Zero-inflated Dirichlet-tree multinomial (ZIDTM) regression.}
\usage{
ZIdtmReg(
  otu.tab,
  X.mean,
  X.disp = NULL,
  X.zero = NULL,
  tree,
  test.type = "Mean",
  index = 1
)
}
\arguments{
\item{otu.tab}{a data frame or matrix containing the count data. Rows of the matrix represent observations and columns are the taxa.}

\item{X.mean}{a data frame or matrix containing the covariates link to the mean parameter. Rows of the matrix represent observations.}

\item{X.disp}{a data frame or matrix containing the covariates link to the dispersion parameter. Rows of the matrix represent observations.}

\item{X.zero}{a data frame or matrix containing the covariates link to the zero prevalance. Rows of the matrix represent observations.}

\item{tree}{the phylogenetic tree}

\item{test.type}{f test.type = "Mean", the function will test for differential mean (Default). If test.type = "Disp", the function will test for differential dispersion. If test.type = "Freq", the function will test for differential presence-absence frequency.}

\item{index}{index position for the interest variable for testing and remain treat as confounders.}
}
\value{
Returns a list, containing testing results of each subtree.
}
\description{
Fit the Zero-inflated Dirichlet-tree multinomial (ZIDTM) regression.
}
\examples{
library(phyloseq)
data("combo.phyloseq.obj")
otu.tab <- t(otu_table(combo.phyloseq.obj))
tree <- phy_tree(combo.phyloseq.obj)
sodium <- sample_data(combo.phyloseq.obj)$sodium
fit <- ZIdtmReg(otu.tab, sodium, tree=tree)
}
\references{
Zhou, C., Zhao, H., and Wang, T.: Transformation and differential abundance analysis of microbiome data incorporating phylogeny. Bioinformatics btab543, https://doi.org/10.1093/bioinformatics/btab543 (2021).
}
