% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eBay_comps.R
\name{eBay_comps}
\alias{eBay_comps}
\title{An empirical Bayes approach for transforming raw counts into relative abundances}
\usage{
eBay_comps(otu_tab, prior, tree = NULL)
}
\arguments{
\item{otu_tab}{a data frame or matrix containing the count data. Rows of the matrix represent observations and columns are the taxa.}

\item{prior}{Dirichlet or Dirichlet-tree prior for the proportions.}

\item{tree}{the phylogenetic tree.}
}
\value{
Returns the estimated microbial compositions by empirical Bayes.
}
\description{
An empirical Bayes approach for transforming raw counts into relative abundances
}
\examples{
library(phyloseq)
otu_tab <- t(combo.phyloseq.obj@otu_table@.Data)
tree <- phy_tree(combo.phyloseq.obj)
eBay.comps <- eBay_comps(otu_tab,prior = "Dirichlet", tree=NULL)
eBay.tree.comps <- eBay_comps(otu_tab,prior = "Dirichlet-tree", tree = tree)
}
\references{
Liu, T., Zhao, H., Wang, T.: An empirical Bayes approach to normalization and differential abundance testing for microbiome data. BMC Bioinformatics 21, 225 (2020).
}
